feat: support negative checkpoint indices for --playback-from-checkpoint
Negative values count from end:
  -1 = second-to-last checkpoint
  -2 = third-to-last checkpoint
  etc.

Resolution formula: actual_index = (total - 1) + n

Changes:
- autorun_from_checkpoint: Option<usize> → Option<i64>
- Add parse_i64_arg() in config.rs (handles =syntax and negatives)
- Add resolve_checkpoint_index() in autorun_state.rs
- Update init_autorun() signature in sdl_eventloop.rs
- 5 new tests covering negative indexing

+++ src/sdl_frontend/sdl_eventloop.rs
@@ -378,7 +378,8 @@ impl SdlEventLoop {
     /// * `rom_path` - Path to the loaded ROM; the `.autorun` file lives alongside it.
     /// * `overwrite` - In `Record` mode: back up any existing file and start fresh.
     /// * `extend` - In `Record` mode: replay from second-to-last checkpoint then continue recording.
-    /// * `from_checkpoint` - In `Playback` mode: start playback from this checkpoint index.
+    /// * `from_checkpoint` - In `Playback` mode: start playback from this checkpoint index
+    ///   (negative = from end, -1 = second-to-last, etc.).
     /// * `nes` - The NES instance; used to restore saved state when starting from a checkpoint.
     pub fn init_autorun(
         &mut self,
@@ -386,7 +387,7 @@ impl SdlEventLoop {
         rom_path: &str,
         overwrite: bool,
         extend: bool,
-        from_checkpoint: Option<usize>,
+        from_checkpoint: Option<i64>,
         nes: &mut Nes,
     ) -> Result<(), String> {
         if mode != AutorunMode::None {
